+++ src/simulator/data.go
@@ -9,7 +9,7 @@ type Chart []data
 
 type data struct {
 	time time.Time
-	value int
+	value float64
 }
 
 func (s data) String() string {
@@ -17,7 +17,7 @@ func (s data) String() string {
 }
 
 func (s data) StringCsv() string {
-	return fmt.Sprintf("%s, %d",s.time.String(),s.value)
+	return fmt.Sprintf("%s, %f",s.time.String(),s.value)
 }
 
 func (s Chart) StringCsv() string {
